requirements.txt
lightgbm==4.6.0
matplotlib==3.10.1
numpy==2.2.3
pandas==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
sentence_transformers==3.3.1
torch==2.5.0+cu118
torchvision==0.20.0+cu118
tqdm==4.66.5
